+++ wasm/src/wasm_html.cpp
@@ -159,6 +159,7 @@ HtmlConfig to_html_config(const emscripten::val &value) {
 
   read(value, "embedImages", config.embed_images);
   read(value, "editable", config.editable);
+  read_enum(value, "editingScope", config.editing_scope);
   read(value, "keyboardNavigation", config.keyboard_navigation);
   read(value, "keyboardShortcuts", config.keyboard_shortcuts);
   read(value, "textDocumentMargin", config.text_document_margin);